Make index.tsx just login
