+++ holium-com/src/consts.ts
@@ -9,6 +9,6 @@ export const constants = {
 export const MOBILE_WIDTH = 560;
 export const DESKTOP_WIDTH = 1480;
 
-export const LOGIN_HREF = 'https://hosting.holium.com/login';
+export const LOGIN_HREF = 'https://hosting.holium.com';
 
-export const GET_REALM_HREF = 'http://hosting.holium.com/get-realm';
+export const GET_REALM_HREF = 'http://hosting.holium.com';
