+++ hosting-holium-com/src/pages/index.tsx
@@ -1,61 +1,122 @@
-import { useEffect } from 'react';
-import type { GetServerSideProps } from 'next/types';
+import { GetServerSideProps } from 'next';
 
-import { GetOnRealmDialog, OnboardingStorage } from '@holium/shared';
+import { useToggle } from '@holium/design-system/util';
+import {
+  ChangePasswordWithTokenModal,
+  ForgotPassword,
+  ForgotPasswordModal,
+  LoginDialog,
+  OnboardingStorage,
+} from '@holium/shared';
+
+import { thirdEarthApi } from 'util/thirdEarthApi';
 
 import { Page } from '../components/Page';
 import { useNavigation } from '../util/useNavigation';
 
 type Props = {
-  email: string;
+  prefilledEmail: string;
+  redirectAfterLogin: string;
 };
 
 export const getServerSideProps: GetServerSideProps = async ({ query }) => {
-  const email = (query.email ?? '') as string;
+  const prefilledEmail = (query.email ?? '') as string;
+  const redirectAfterLogin = (query.redirect_url ?? '') as string;
 
   return {
     props: {
-      email,
+      prefilledEmail,
+      redirectAfterLogin,
     },
   };
 };
 
-export default function GetOnRealm({ email }: Props) {
+export default function Login({ prefilledEmail, redirectAfterLogin }: Props) {
   const { goToPage } = useNavigation();
 
-  const onBack = () => {
-    window.location.href = 'https://holium.com';
+  const forgotPassword = useToggle(false);
+  const resetPassword = useToggle(false);
+
+  const onBack = () => goToPage('/');
+
+  const onForgotPassword = async (email: string) => {
+    const response = await thirdEarthApi.forgotPassword(email);
+
+    if (response) {
+      OnboardingStorage.set({ email });
+
+      forgotPassword.toggleOff();
+      resetPassword.toggleOn();
+
+      return true;
+    } else {
+      return false;
+    }
   };
 
-  const onUploadPier = () => {
-    return goToPage('/create-account', {
-      haha: 'true',
-      product_type: 'byop-p',
-    });
+  const onResetPassword = async (token: string, password: string) => {
+    const response = await thirdEarthApi.resetPassword(token, password);
+
+    if (response?.token) {
+      OnboardingStorage.set({ token: response.token });
+
+      resetPassword.toggleOff();
+
+      return true;
+    } else {
+      return false;
+    }
   };
 
-  const onPurchaseId = async () => {
-    return goToPage('/create-account', {
-      haha: 'true',
-    });
-  };
+  const onLogin = async (email: string, password: string) => {
+    const response = await thirdEarthApi.login(email, password, true);
 
-  const onAlreadyHaveAccount = () => {
-    return goToPage('/login');
-  };
+    if (!response.token || !response.email) {
+      return false;
+    } else {
+      OnboardingStorage.set({
+        token: response.token,
+        email: response.email,
+        clientSideEncryptionKey: response.client_side_encryption_key,
+      });
+    }
 
-  useEffect(() => {
-    OnboardingStorage.set({ email });
-  }, [email]);
+    if (redirectAfterLogin) goToPage(redirectAfterLogin as any);
+    else {
+      const ships = await thirdEarthApi.getUserShips(response.token);
+
+      if (ships.length) {
+        // Ships imply Realm access.
+        goToPage('/account');
+      } else if (response.client_side_encryption_key) {
+        // CSEK without ships still implies Realm access.
+        goToPage('/account/download-realm');
+      } else {
+        // No ships or CSEK implies no Realm access.
+        goToPage('/account/get-realm');
+      }
+    }
+
+    return true;
+  };
 
   return (
-    <Page title="Get on Realm">
-      <GetOnRealmDialog
-        onUploadPier={onUploadPier}
-        onPurchaseId={onPurchaseId}
-        // Email query parameter means they're coming from the landing page.
-        onBack={email ? onBack : undefined}
-        onAlreadyHaveAccount={onAlreadyHaveAccount}
+    <Page title="Login">
+      <ForgotPasswordModal
+        isOpen={forgotPassword.isOn}
+        onDismiss={forgotPassword.toggleOff}
+        onSubmit={onForgotPassword}
+      />
+      <ChangePasswordWithTokenModal
+        isOpen={resetPassword.isOn}
+        onDismiss={resetPassword.toggleOff}
+        onSubmit={onResetPassword}
+      />
+      <LoginDialog
+        prefilledEmail={prefilledEmail}
+        footer={<ForgotPassword onClick={() => forgotPassword.toggleOn()} />}
+        onBack={onBack}
+        onLogin={onLogin}
       />
     </Page>
   );
