+++ shared/src/onboarding/onboarding.ts
@@ -17,8 +17,6 @@ export { SubmitButton } from './components/SubmitButton';
 export { TermsDisclaimer } from './components/TermsDisclaimer';
 export { AccountCustomDomainDialog } from './dialogs/AccountCustomDomain/AccountCustomDomainDialog';
 export { AccountCustomDomainDialogBody } from './dialogs/AccountCustomDomain/AccountCustomDomainDialogBody';
-export { AccountDownloadRealmDialog } from './dialogs/AccountDownloadRealm/AccountDownloadRealmDialog';
-export { AccountGetRealmDialog } from './dialogs/AccountGetRealm/AccountGetRealmDialog';
 export { AccountHostingDialog } from './dialogs/AccountHosting/AccountHostingDialog';
 export { AccountHostingDialogBody } from './dialogs/AccountHosting/AccountHostingDialogBody';
 export { AccountSelfHostingDialogBody } from './dialogs/AccountSelfHosting/AccountSelfHostingDialogBody';
